vehicle/viss/0300-subscribe-filter-range-success.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>
[0300: filter range]
</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">
<link rel="stylesheet" type="text/css" href="./style-general.css">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-config.js"></script>
<script src="./vehicle-test-helper.js"></script>

<div id="test-title">
[0300: filter range]
</div>
<br>
<div id="test-assertion">
<b>Test assertion:</b><br>
If the client specifies the "Range" of the signal,<br>
 the client receives data when the value satifies the conditions.<br>
</div>
<div id="scenario">
<br>
<b>Test Scenario:</b><br>
1. Execute Subscribe method to a data path with specifing 'range' as filter.<br>
2. Client will receive 'SubscriptionNotification' only when the value satisfies the conditions.<br>
<br>
<b>Expected Result:</b><br>
- Receive 'SubscriptionNotification' only when its value satisfies the 'range' conditions <br>
</div>

<div id="result"></div>
<br>
<div id="log"></div>

<script>

// phase
var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

var reqId = "reqid-0300";
var path = SUBSCRIBE_STANDARD_PATH;
var phase = PH_START;
var subId = null;

var notifCount = 0;
var notif = [];
var notif_timer_started = false;

var intervalTime   = INTERVAL_TIME
var intervalMargin = INTERVAL_MARGIN;
var range_above    = RANGE_ABOVE;
var range_below    = RANGE_BELOW;
var minChangeVal   = MINCHANGE_VAL;

var filter_interval = '{"interval":' + intervalTime + '}';
var filter_range = '{"range": {"above":' + range_above + ', "below":'+range_below + '}}';
var filter_minChange = '{"range": {"minChange":' + minChangeVal + '}';

var t = async_test("Subscribe test: subscribe with filter=Range");
var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

vehicle.onopen = function() {

  // process messages from the server
  vehicle.onmessage = function(event) {
    addLogMessage("Response: json="+sub_req);
    var ts = getTimestamp();
    var msg = JSON.parse(event.data);

    if (phase == PH_START) {
      // nothing to do
    } else if (phase == PH_SENT_SUBSCRIBE) {
      t.step_func(function() {
        // wait for subscribe response
        if (msg.action !== "subscribe") {
          addLogMessage("PH_SENT_SUBSCRIBE: Ignore messages except for Subscribe responses.");
          addLogMessage("  msg = " + JSON.stringify(msg));
          return;
        }
        if (isSubscribeSuccessResponse(reqId, msg)) {
          addLogMessage("subscribeSuccessResponse received. subscribeId = " + msg.subscriptionId);
          subId = msg.subscriptionId
          phase = PH_SUBSCRIBED;
        } else if (isSubscribeErrorResponse(reqId, msg)) {
          helper_terminate_failure("subscribeErrorResponse received. test failed.");
          phase = PH_END;
        }
      })();
    } else if (phase == PH_SUBSCRIBED) {
      t.step_func(function() {
        // wait subscriptionNotification
        if (notif_timer_started === false) {
          notif_timer_started = true;
          // wait for subscriotionNotification for fixed time
          t.step_timeout(function() {
            // When wait time over, do unsubscribe
            addLogMessage("Send unsubscribe request");
            var unsub_req ='{"action":"unsubscribe"'
                          +' ,"subscriptionId":"'+subId+'"'
                          +' ,"requestId":"'+reqId+'"'
                          +'}';
            addLogMessage("unsubscribe req: json="+unsub_req);
            vehicle.send(unsub_req);
            phase = PH_SENT_UNSUBSCRIBE;
          }, TIME_NOTIF_WAIT
          );
        }

        if (msg.action !== "subscription") {
          addLogMessage("PH_SUBSCRIBED: Ignore messages except for Subscription notification.");
          addLogMessage("msg = " + JSON.stringify(msg));
          return;
        }
        if (isSubscriptionNotificationResponse(subId, msg)) {
          notif[notifCount] = {};
          notif[notifCount].json = msg;
          notif[notifCount].timestamp = ts;
          addLogMessage("SubscriptionNotification received. value = " + msg.value + ", timestamp = " + ts);
          notifCount++;
        } else if (isSubscriptionNotificationErrorResponse(subId, msg)) {
          helper_terminate_failure("SubscriptionNotificationError received. Abort test.");
          phase = PH_END;
        }
      })();
    } else if (phase == PH_SENT_UNSUBSCRIBE) {
      t.step_func(function() {
        // wait for unsubscribe response
        if (msg.action !== "unsubscribe") {
          addLogMessage("PH_SENT_UNSUBSCRIBE: Ignore messages except for Unsubscribe response.");
          addLogMessage("  msg = " + JSON.stringify(msg));
          return;
        }
        var res = isUnsubscribeSuccessResponse(reqId, subId, msg);
        assert_true(res, "True if unsubscribeSuccessResponse received.");
        if (res) {
          // unsubscibe success
          addLogMessage("subscribe/unsubscribe successed");
        } else if (isUnsubscribeErrorResponse(reqId, subId, msg)) {
          // unsubscribe error
          // Test has been success already. Terminate anyway.
          addLogMessage("unsubscribeErrorResponse received");
        }
        phase = PH_UNSUBSCRIBED;
        var flgSuccess = true;
        // check the notification result
        for (var i=0; i<notifCount; i++) {
          var res = verifyRange(notif[i], range_above, range_below );
          if (res == false)
            flgSuccess = false;
        }
        if (flgSuccess == true) {
          helper_terminate_success("All the notification satisfied filter condition");
        } else {
          helper_terminate_failure("Some notification didn't satisfy filter condition");
        }

      })();
    } else if (phase == PH_UNSUBSCRIBED) {
      //nothing to do
    }
  }

  // in case server doesn't return response
  t.step_timeout(function() {
    if (phase < PH_SUBSCRIBED) {_
      helper_terminate_failure("Finish by timeout");
    }
  }, TIME_OUT_TIME
  );

  addLogMessage("Send subscribe request: path= "+path+", filter= " + filter_range+ "<br>");
  var sub_req = '{"action":"subscribe","path":"'+ path +'"'
                +',"requestId":"'+reqId+'"'
                +',"filters":'+filter_range+'}';
  addLogMessage("subscribe req: json="+sub_req);
  vehicle.send(sub_req);

  phase = PH_SENT_SUBSCRIBE;
}

// verify Range
// - supposing '_above' and '_below' are number or null or undefined
function verifyRange(_notif, _above, _below) {
  var val = _notif.json.value;
  if (_below && val > _below) {
    addLogFailure("value: "+ val + " > Range_below:"+ range_below);
    return false;
  } else
  if ( _above && val < _above) {
    addLogFailure("Range_above:"+ range_above +" > value:"+ val);
    return false;
  } else {
    addLogSuccess("Range_below:"+ _below +" >= value:"+ val +" >= Range_above:"+ _above);
    return true;
  }
}

</script>
</html>
